Buffered writes for metadata updates
The write buffer builds up a buffer to write to an fd, in-memory, so that
the entire buffer can be written at once, in as few writes as possible.

The buffer grows as needed depending on what is being written, starting
at an initial chunk size, and growing in chunk-sized increments.

If a short write occurs, write(2) is continued to be called with the
remaining length of the buffer until successful. If a write(2) error
occurs, then the write is aborted, and error is returned.