calc_array_size: correctly factor in chunk size
I had updated calc_array_size some time ago to do a "pretty good" job at
computing the true size of a RAID-10 volume with an odd number of disks.  It
worked pretty well -- until we encountered drives of an odd size that, when
combined into a 5-disk RAID-10, produced an array with a size the kernel
rejected.

Initially, I tried simply aligning the array length to the chunk size.  But,
after reviewing the kernel code, we found this didn't accurately capture what
needed to happen for arrays with a "far" layout.  We ended up with something
very similar to the way the kernel computes the size of the array.