Ensure disk recovery starts for container re-add
A re-add must recover the added disk from the beginning. This ensures
that the entire bitmap is consulted when resyncing the added disk. In
the re-add case, set recovery_start to 0 for the disk as part of the
sysfs dance (the order between insync, recovery_start is critical).
mdmon here does the real work, by passing "resume" into the sysfs
function to do the re-add.

Also, fix a bug with raid_disk reporting. This fixes the log message
issue where a re-added disk had -1 printed, even though it had been
assigned a correct raid_disk.