Set events_cleared in bitmap file to ensure assembly bitmap recovery
The kernel checks array 'events' with bitmap 'events_cleared'. If they
are the same, then the bits loaded in the bitmap on assembly are
discarded, even on a degraded array. This is a specific bug for external
(ie: ddf) metadata, as external metadata has no interface and no ability
to set the array 'events' to keep the kernel informed and in-sync.

Workaround this lack of 'events' tracking, so that bitmap load *flags*
the bits in the bitmap as NEEDED for a degraded array, by writing
a bogus non-zero 'events_cleared' value to the bitmap file during mdadm
assembly. As external metadata can only be used with a bitmap file
(internal bitmaps not supported), and because the 'events_cleared' value
only has to be *different* during bitmap load in the kernel, any
non-zero value works. Once the bitmap has been loaded and flushed, the
bitmap superblock is updated in the kernel resetting it to 0. But the
bits have been loaded correctly, and are not cleared.

A drive re-add will use the bitmap for recovery correctly.