Re-add possible only if disk is MD_DISK_SYNC.
Fallback on re-add for a "rebuilding" disk, to treat it as a new "add".
This ensures that if rebuild/recovery/initial-sync did not complete,
that we do not perform bitmap recovery on the added disk. The add gets
treated directly as a new "add", and the disk recovery starts over.