Probe device before add in mdadm and mdmon
Before adding a device to a container, probe the device to make sure it
is online. This does a simple read probe, the same as our mdmon device
probe.

We probe _before_ doing a re-add/add in mdadm. Also, if the device goes
offline after mdadm, but before mdmon wakes up and adds the disk to the
container, we probe *again* in mdmon before updating the metadata.

Without the probe, we attempt to read the metadata for the device (and
cannot do so because it is offline) and then treat the disk as a "new"
device being added to the array. mdmon wakes up, adds the disk to the
container, updates everyone elses metatadata. But the disk added is
immediately faulty because it is offline. So mdmon then removes it. The
removal triggers an mdvote "member" event increment, which then triggers
the offline device as no longer a valid member of the array. When it
does eventually come back online, even if no new replacement was added
to the array in the meanwhile, this otherwise valid disk cannot be
re-added, and we must wait until time2repair for the volume when an
unmap of the device and an add as new takes place. So, we lose the
ability to bitmap recover this offline device and must take a full
resync.

This probe addresses this issue, so we don't make forward progress in
the array membership when we shouldn't.