Skip metadata writes to "failed" devices
For a device going away and coming back, it can race with metadata
updates. Because additional metadata updates are going to come in, for
write pending transitions for example, a device which failed a metadata
update because it was "offline", can later succeed when it comes back.
Before it has a chance to be removed from the container.

If this happens, the formerly "offline" device gets its *own* metadata
marked as "failed" for itself. This prevents a future re-add, as
"failed" devices are not allowed to be "re-add"s. There is no reason to
flush out a "failed" state to the actual device that is failed. We have
mdvote for array membership checks, and the remaining members have the
"failed" state for that drive, themselves. We likely *want* a re-add, in
the cases that we can have it.

Implement this by skipping the metadata writes for DDF_Failed.