Fix two memory leaks and an uninitialized read
The ddf_probe_device_fd() failed to respect the 'static' nature of the
posix_memaligned buffer, after the last change. Move back to only
allocating the buffer once and re-using it. Also fix an unitialized read
for the printf in here.

sysfs_read() returns an allocated mdinfo, but most (all?) cases did not
free it. Have the callers free the memory using sysfs_free().