Always update array state on membership changes
A specific sequence of disk failures and a race with the kernel
signaling to remove a disk from the array, can cause an array to lose
a member disk but not have the metadata persistently mark the array as
degraded.

The sequence is:

  * disk goes away (dev entry missing / gone)
  * disk is removed from container, tracking structures zeroed out
  * kernel says disk is "faulty".
  * attempt to mark the faulty disk as "faulty"
  * skip it ; disk doesn't exist ; array metadata never marked degraded

Later, due to this missing degraded, when a new replacement disk is
*added*, the array is not degraded, and this new disk is not marked
rebuilding. However, the new disk *must* be marked rebuilding. And the
array *must* be marked degraded so additional checks/logic can be
performed.

Ensure correct array state by updating it *always*, even if the original
disk goes away. And also check it specifically again on disk *add* and
disk *remove*. This catches and closes other races with updating the
array to degraded during reboots/failover.

Additionally, improve various log messages around "state" and
"init_state" for better debuggability in log messages.