treat missing disks as failed
A "missing" disk is a "failed" disk. When setting a disk to "missing",
ensure that it is also set to "failed". When checking for "failed",
also check for "missing" as an equivalence.

Treat "missing" disks as "failed" and "not working", for calculating
array state. A "missing" disk cannot be a "working" disk.