DDF: take MD_DEVNAME directly from name of array not 'map' file
We have seen the map file get corrupted under load (missing devname). In
that case the array assembles, but MD_DEVNAME never gets set as the udev
rule sets /dev/md/DEVNAME from the MD_DEVNAME exposed in --detail
--export.

For DDF we can take this directly from the info->name (array name) of
the subarray. We do not set a homehost, not does DDF expose it anyway.
This ensures any map corruption we can ride out.