Check for errors when loading the metadata
During load_super_ddf_all() reading of headers or global data can fail.
However, the error code was ignored. Later, when the super->active
structure was dereferenced, a segfault occurred as super->active was
never set due to the loading of header failure.

The loading of header failed, because the metadata was "open" for write,
which is possible when the metadata was being updated.

This meant, both the primary and the secondary metadata sections were
"open" at time of read. This is possible, because the order in which the
sections are written, and the order in which the sections are read, are
the same. Alleviate this by flipping the write order.