Use separate 'replace' tracker variable in the per-disk mdinfo structure
mdmon uses this variable to determine that a disk has been marked as
"wanting replacement". As there are races in the kernel state, we must
keep track if the kernel has ever marked a disk as "want_replacement",
so that when that disk is marked "faulty" by the kernel, we know to
"replace" it, and not just "fail it".

Usually, "want_replacement" and "faulty" are marked at the same time
when the spare synchronization is complete, but there is a race where
"want_replacement" is no longer marked in the kernel when mdmon wakes up
for process, but "faulty" is currently marked. We need to track the
"faulty,want_replacement" combo state ourselves to be sure.

Additionally, update a bunch of debug log messages for consistency.